+++ ChangeLog.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 - Filesystem::unlink()
 - Filesystem::rmdir()
 - Filesystem::mkdir()
+- Filesystem::mknod()
 - Filesystem::chown()
 - Filesystem::chmod()
 - Filesystem::utime()